converza-pipeline/pyproject.toml
[project]
name = "converza-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.0.6",
    "boto3>=1.40.37",
    "dash>=3.2.0",
    "faker>=37.8.0",
    "minio>=7.2.16",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=1.4.54",
]
